在室状況を表すstatusコンポーネントを作成し、topページに導入しました
