--- src/app/admin/login/page.tsx
@@ -1,70 +1,0 @@
-"use client"
-
-import type React from "react"
-
-import { useState } from "react"
-import { useRouter } from "next/navigation"
-import { Button } from "@/components/ui/button"
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { adminLogin } from "@/lib/actions"
-
-export default function AdminLoginPage() {
-  const router = useRouter()
-  const [email, setEmail] = useState("")
-  const [password, setPassword] = useState("")
-  const [error, setError] = useState("")
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setError("")
-
-    try {
-      const result = await adminLogin({ email, password })
-      if (result.success) {
-        router.push("/admin/dashboard")
-      } else {
-        setError("ログインに失敗しました。管理者アカウントの認証情報を確認してください。")
-      }
-    } catch (err) {
-      setError("エラーが発生しました。もう一度お試しください。")
-    }
-  }
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-muted/40 px-4">
-      <Card className="w-full max-w-md">
-        <CardHeader>
-          <CardTitle className="text-2xl">管理者ログイン</CardTitle>
-          <CardDescription>研究室管理者としてログインしてください</CardDescription>
-        </CardHeader>
-        <form onSubmit={handleSubmit}>
-          <CardContent className="space-y-4">
-            {error && <div className="bg-destructive/10 text-destructive text-sm p-3 rounded-md">{error}</div>}
-            <div className="space-y-2">
-              <Label htmlFor="email">管理者メールアドレス</Label>
-              <Input id="email" type="email" value={email} onChange={(e) => setEmail(e.target.value)} required />
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="password">パスワード</Label>
-              <Input
-                id="password"
-                type="password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                required
-              />
-            </div>
-          </CardContent>
-          <CardFooter>
-            <Button type="submit" className="w-full">
-              管理者ログイン
-            </Button>
-          </CardFooter>
-        </form>
-      </Card>
-    </div>
-  )
-}
-
